change f64 to f32 quality scores

+++ src/main.rs
@@ -29,7 +29,7 @@ fn main() -> Result<(), Error> {
  
     let min_length: u64 = cli.value_of("MINLEN").unwrap_or("0").parse().unwrap();
     let max_length: u64 = cli.value_of("MAXLEN").unwrap_or("0").parse().unwrap();
-    let min_quality: f64 = cli.value_of("QUALITY").unwrap_or("0").parse().unwrap();
+    let min_quality: f32 = cli.value_of("QUALITY").unwrap_or("0").parse().unwrap();
     let crab: bool = cli.is_present("CRAB");
 
         
@@ -104,7 +104,7 @@ fn crabcast(min_length: u64, max_length: u64, min_quality: f32) -> Result<(u64, 
 
         let quality_values: Vec<u8> = record.qual().to_vec();
         let mean_error: f32 = get_mean_error(&quality_values);
-        let mean_quality: f32 = -10f32*log10(mean_error as f32);
+        let mean_quality: f32 = (-10f64*log10(mean_error as f64)) as f32;
 
         let seqlen = record.seq().len() as u64;
                 
@@ -148,7 +148,7 @@ fn needlecast_filter(min_length: u64, max_length: u64, min_quality: f32) -> Resu
         // Quality scores present:
         if let Some(qual) = seqrec.qual() {
             let mean_error: f32 = get_mean_error(&qual);
-            let mean_quality: f32 = -10f32*log10(mean_error as f32);
+            let mean_quality: f32 = (-10f64*log10(mean_error as f64)) as f32;
             // Fastq filter
             if seqlen >= min_length && mean_quality >= min_quality && seqlen <= max_length {
                 reads += 1;
@@ -192,7 +192,7 @@ fn needlecast_stats() -> Result<(u64, u64, Vec<u64>, Vec<f32>), Error> {
         // Quality scores:
         if let Some(qual) = seqrec.qual() {
             let mean_error: f32 = get_mean_error(&qual);
-            let mean_quality: f32 = -10f32*log10(mean_error as f32);
+            let mean_quality: f32 = (-10f64*log10(mean_error as f64)) as f32;
             read_qualities.push(mean_quality);
         }
